Employee Schedule
Allows for creation of employee schedule. As of now there is yet no update or delete, that will be added later. The next thing to do for now is to create the view for the schedule.

Happy New Year to me. Here's to me pushing through and do this piece by piece up until the end. Cheers!